feat(ebay_watcher): async client, BidStatus enum, retry/backoff, TUI improvements
- Refactor eBayClient trait to async with mock gated behind cfg(test)/feature flag
- Add BidStatus enum for type-safe status tracking in models
- Add is_transient() error detection and exponential backoff retry in sync engine
- Refactor run_daemon to accept Config directly, use async everywhere
- Write PID file atomically (temp + rename) to prevent partial reads
- Use nix::signal instead of raw libc::kill in TUI for proper error handling
- Add snipe cancellation (x key) and filtered_items caching in TUI
- Pull deals max_total_price from config instead of hardcoded value
- Replace inline SQL with db module helpers in TUI
- Add unit tests for error, models, sync, scheduler, and TUI
- Remove hello_world skill

+++ internal_projects/ebay_watcher/ebay_watcher_tui/src/ui/dashboard.rs
@@ -11,7 +11,7 @@ use std::path::Path;
 use ebay_watcher_core::models::{Item, Alert, OAuthToken};
 
 pub struct DashboardState {
-    pub daemon_pid: Option<i32>,
+    pub daemon_pid: Option<u32>,
     pub daemon_active: bool,
     pub oauth_token: Option<OAuthToken>,
     pub total_items: usize,
@@ -28,10 +28,11 @@ impl DashboardState {
         let mut daemon_active = false;
         if Path::new(pid_file).exists() {
             if let Ok(pid_str) = fs::read_to_string(pid_file) {
-                if let Ok(pid) = pid_str.trim().parse::<i32>() {
+                if let Ok(pid) = pid_str.trim().parse::<u32>() {
                     daemon_pid = Some(pid);
-                    // Check if process is alive (using kill 0)
-                    daemon_active = unsafe { libc::kill(pid, 0) == 0 };
+                    // Display-only liveness probe via signal 0. This is a snapshot;
+                    // callers must NOT rely on it for security-sensitive decisions (TOCTOU).
+                    daemon_active = unsafe { libc::kill(pid as i32, 0) == 0 };
                 }
             }
         }
